removed extra info from problemConfig.
Initial state is generated form the content, and the template url was
moved to the problem ui module (because it's the same for every problem
of a given category).

Also removed the guards for whether the category is registered and just
throw an error if it's not, because silently doing nothing if no
category is registered is never what we want.