Fixed If/Else needing a def
Currently only the TPL for If/Else has been changed so that it prints out the right prompts for code without function calls. It has been kept so that all old legacy examples work exactly as they did before. Doing functionless examples for the other practicum modes should work as long as the TPL is changed.